Add player health
The player now has health and cacti can damage the player. The player health script has functions for damaging the player, healing the player, and setting the maximum health. There is also a player health bar which listens for changes in the player's health and updates accordingly.

Also, the Objects folder has been renamed to Obstacles.