Make rooms generate all at once
Rooms are now generated all at once rather than one tile at a time. This is more efficient, and it should make it easier to generate the interiors of the rooms if we do it all in the same scene and use random interiors.

Also, the way cactus biomes generate has been slightly changed.